# [Merged by Bors] - Frustum Culling (for Sprites)