--- a/crates/bevy_render/src/camera/active_cameras.rs
+++ b/crates/bevy_render/src/camera/active_cameras.rs
@@ -21,11 +21,13 @@ pub struct ActiveCameras {
 
 impl ActiveCameras {
     pub fn add(&mut self, name: &str) {
-        self.cameras
-            .insert(name.to_string(), ActiveCamera {
+        self.cameras.insert(
+            name.to_string(),
+            ActiveCamera {
                 name: name.to_string(),
                 ..Default::default()
-            });
+            },
+        );
     }
 
     pub fn get(&self, name: &str) -> Option<&ActiveCamera> {
@@ -36,11 +38,11 @@ impl ActiveCameras {
         self.cameras.get_mut(name)
     }
 
-    pub fn iter(&self) -> impl Iterator<Item=&ActiveCamera> {
+    pub fn iter(&self) -> impl Iterator<Item = &ActiveCamera> {
         self.cameras.values()
     }
 
-    pub fn iter_mut(&mut self) -> impl Iterator<Item=&mut ActiveCamera> {
+    pub fn iter_mut(&mut self) -> impl Iterator<Item = &mut ActiveCamera> {
         self.cameras.values_mut()
     }
 }