--- a/crates/bevy_render/src/draw.rs
+++ b/crates/bevy_render/src/draw.rs
@@ -70,7 +70,7 @@ impl Default for Visible {
 /// Any Sprite or AtlasTextureSprite will have this removed if they are outside the camera frustum and thus not be rendered.
 #[derive(Debug, Default, Clone, Reflect)]
 #[reflect(Component)]
-pub struct WithinFrustum;
+pub struct OutsideFrustum;
 
 /// A component that indicates how to draw an entity.
 #[derive(Debug, Clone, Reflect)]