--- a/crates/bevy_render/src/pipeline/render_pipelines.rs
+++ b/crates/bevy_render/src/pipeline/render_pipelines.rs
@@ -1,13 +1,13 @@
 use super::{PipelineDescriptor, PipelineSpecialization};
 use crate::{
-    draw::{Draw, DrawContext, WithinFrustum},
+    draw::{Draw, DrawContext, OutsideFrustum},
     mesh::{Indices, Mesh},
     prelude::{Msaa, Visible},
     renderer::RenderResourceBindings,
 };
 use bevy_asset::{Assets, Handle};
 use bevy_ecs::{
-    query::With,
+    query::Without,
     reflect::ReflectComponent,
     system::{Query, Res, ResMut},
 };
@@ -89,7 +89,7 @@ pub fn draw_render_pipelines_system(
     meshes: Res<Assets<Mesh>>,
     mut query: Query<
         (&mut Draw, &mut RenderPipelines, &Handle<Mesh>, &Visible),
-        With<WithinFrustum>,
+        Without<OutsideFrustum>,
     >,
 ) {
     for (mut draw, mut render_pipelines, mesh_handle, visible) in query.iter_mut() {