--- a/crates/bevy_sprite/src/frustum_culling.rs
+++ b/crates/bevy_sprite/src/frustum_culling.rs
@@ -1,7 +1,10 @@
 use bevy_asset::{Assets, Handle};
 use bevy_ecs::prelude::{Commands, Entity, Query, Res, With};
 use bevy_math::Vec2;
-use bevy_render::{camera::Camera, draw::OutsideFrustum};
+use bevy_render::{
+    camera::{ActiveCameras, Camera},
+    draw::OutsideFrustum,
+};
 use bevy_transform::components::Transform;
 use bevy_window::Windows;
 
@@ -25,10 +28,11 @@ impl Rect {
     }
 }
 
-pub fn sprites(
+pub fn sprite_frustum_culling_system(
     mut commands: Commands,
     windows: Res<Windows>,
-    cameras: Query<&Transform, With<Camera>>,
+    active_cameras: Res<ActiveCameras>,
+    camera_transforms: Query<&Transform, With<Camera>>,
     culled_sprites: Query<&OutsideFrustum, With<Sprite>>,
     sprites: Query<(Entity, &Transform, &Sprite)>,
 ) {
@@ -38,36 +42,39 @@ pub fn sprites(
         return;
     };
 
-    for camera_transform in cameras.iter() {
-        let camera_size = window_size * camera_transform.scale.truncate();
+    for active_camera_entity in active_cameras.iter().filter_map(|a| a.entity) {
+        if let Ok(camera_transform) = camera_transforms.get(active_camera_entity) {
+            let camera_size = window_size * camera_transform.scale.truncate();
 
-        let rect = Rect {
-            position: camera_transform.translation.truncate(),
-            size: camera_size,
-        };
-
-        for (entity, drawable_transform, sprite) in sprites.iter() {
-            let sprite_rect = Rect {
-                position: drawable_transform.translation.truncate(),
-                size: sprite.size,
+            let rect = Rect {
+                position: camera_transform.translation.truncate(),
+                size: camera_size,
             };
 
-            if rect.is_intersecting(sprite_rect) {
-                if culled_sprites.get(entity).is_ok() {
-                    commands.entity(entity).remove::<OutsideFrustum>();
+            for (entity, drawable_transform, sprite) in sprites.iter() {
+                let sprite_rect = Rect {
+                    position: drawable_transform.translation.truncate(),
+                    size: sprite.size,
+                };
+
+                if rect.is_intersecting(sprite_rect) {
+                    if culled_sprites.get(entity).is_ok() {
+                        commands.entity(entity).remove::<OutsideFrustum>();
+                    }
+                } else if culled_sprites.get(entity).is_err() {
+                    commands.entity(entity).insert(OutsideFrustum);
                 }
-            } else if culled_sprites.get(entity).is_err() {
-                commands.entity(entity).insert(OutsideFrustum);
             }
         }
     }
 }
 
-pub fn atlases(
+pub fn atlas_frustum_culling_system(
     mut commands: Commands,
     windows: Res<Windows>,
+    active_cameras: Res<ActiveCameras>,
     textures: Res<Assets<TextureAtlas>>,
-    cameras: Query<&Transform, With<Camera>>,
+    camera_transforms: Query<&Transform, With<Camera>>,
     culled_sprites: Query<&OutsideFrustum, With<TextureAtlasSprite>>,
     sprites: Query<(
         Entity,
@@ -79,30 +86,32 @@ pub fn atlases(
     let window = windows.get_primary().unwrap();
     let window_size = Vec2::new(window.width(), window.height());
 
-    for camera_transform in cameras.iter() {
-        let camera_size = window_size * camera_transform.scale.truncate();
+    for active_camera_entity in active_cameras.iter().filter_map(|a| a.entity) {
+        if let Ok(camera_transform) = camera_transforms.get(active_camera_entity) {
+            let camera_size = window_size * camera_transform.scale.truncate();
 
-        let rect = Rect {
-            position: camera_transform.translation.truncate(),
-            size: camera_size,
-        };
+            let rect = Rect {
+                position: camera_transform.translation.truncate(),
+                size: camera_size,
+            };
 
-        for (entity, drawable_transform, sprite, atlas_handle) in sprites.iter() {
-            if let Some(atlas) = textures.get(atlas_handle) {
-                if let Some(sprite) = atlas.textures.get(sprite.index as usize) {
-                    let size = Vec2::new(sprite.width(), sprite.height());
+            for (entity, drawable_transform, sprite, atlas_handle) in sprites.iter() {
+                if let Some(atlas) = textures.get(atlas_handle) {
+                    if let Some(sprite) = atlas.textures.get(sprite.index as usize) {
+                        let size = Vec2::new(sprite.width(), sprite.height());
 
-                    let sprite_rect = Rect {
-                        position: drawable_transform.translation.truncate(),
-                        size,
-                    };
+                        let sprite_rect = Rect {
+                            position: drawable_transform.translation.truncate(),
+                            size,
+                        };
 
-                    if rect.is_intersecting(sprite_rect) {
-                        if culled_sprites.get(entity).is_ok() {
-                            commands.entity(entity).remove::<OutsideFrustum>();
+                        if rect.is_intersecting(sprite_rect) {
+                            if culled_sprites.get(entity).is_ok() {
+                                commands.entity(entity).remove::<OutsideFrustum>();
+                            }
+                        } else if culled_sprites.get(entity).is_err() {
+                            commands.entity(entity).insert(OutsideFrustum);
                         }
-                    } else if culled_sprites.get(entity).is_err() {
-                        commands.entity(entity).insert(OutsideFrustum);
                     }
                 }
             }