--- a/crates/bevy_sprite/src/lib.rs
+++ b/crates/bevy_sprite/src/lib.rs
@@ -3,6 +3,7 @@ pub mod entity;
 
 mod color_material;
 mod dynamic_texture_atlas_builder;
+mod frustum_culling;
 mod rect;
 mod render;
 mod sprite;
@@ -26,10 +27,14 @@ pub use texture_atlas_builder::*;
 
 use bevy_app::prelude::*;
 use bevy_asset::{AddAsset, Assets, Handle, HandleUntyped};
-use bevy_ecs::system::IntoSystem;
+use bevy_ecs::{
+    component::{ComponentDescriptor, StorageType},
+    system::IntoSystem,
+};
 use bevy_math::Vec2;
 use bevy_reflect::TypeUuid;
 use bevy_render::{
+    draw::WithinFrustum,
     mesh::{shape, Mesh},
     pipeline::PipelineDescriptor,
     render_graph::RenderGraph,
@@ -50,6 +55,9 @@ impl Plugin for SpritePlugin {
             .register_type::<Sprite>()
             .register_type::<SpriteResizeMode>()
             .add_system_to_stage(CoreStage::PostUpdate, sprite_system.system())
+            .add_system_to_stage(CoreStage::PostUpdate, frustum_culling::sprites.system())
+            .add_system_to_stage(CoreStage::PostUpdate, frustum_culling::atlases.system())
+            .add_system_to_stage(CoreStage::PostUpdate, frustum_culling::other.system())
             .add_system_to_stage(
                 CoreStage::PostUpdate,
                 material_texture_detection_system.system(),
@@ -59,16 +67,25 @@ impl Plugin for SpritePlugin {
                 asset_shader_defs_system::<ColorMaterial>.system(),
             );
 
-        let world = app.world_mut().cell();
-        let mut render_graph = world.get_resource_mut::<RenderGraph>().unwrap();
-        let mut pipelines = world
+        let world = app.world_mut();
+        world
+            .register_component(ComponentDescriptor::new::<WithinFrustum>(
+                StorageType::SparseSet,
+            ))
+            .unwrap();
+
+        let world_cell = world.cell();
+        let mut render_graph = world_cell.get_resource_mut::<RenderGraph>().unwrap();
+        let mut pipelines = world_cell
             .get_resource_mut::<Assets<PipelineDescriptor>>()
             .unwrap();
-        let mut shaders = world.get_resource_mut::<Assets<Shader>>().unwrap();
+        let mut shaders = world_cell.get_resource_mut::<Assets<Shader>>().unwrap();
         crate::render::add_sprite_graph(&mut render_graph, &mut pipelines, &mut shaders);
 
-        let mut meshes = world.get_resource_mut::<Assets<Mesh>>().unwrap();
-        let mut color_materials = world.get_resource_mut::<Assets<ColorMaterial>>().unwrap();
+        let mut meshes = world_cell.get_resource_mut::<Assets<Mesh>>().unwrap();
+        let mut color_materials = world_cell
+            .get_resource_mut::<Assets<ColorMaterial>>()
+            .unwrap();
         color_materials.set_untracked(Handle::<ColorMaterial>::default(), ColorMaterial::default());
         meshes.set_untracked(
             QUAD_HANDLE,