… crates/bevy_sprite/Cargo.toml
… crates/bevy_sprite/src/sprite.rs
… crates/bevy_text/src/text2d.rs
… crates/bevy_ui/src/widget/text.rs