@Byteron 2021-02-22
^I broke something when fixing clippy things. Fixed that and squashed commits.
(Hence the force push)