@Byteron 2021-02-22
> It would be nice to have a benchmark to spot regressions and toss in the patch notes, but I think this PR is good to go and would be a really nice easy win for 0.5.

Where would I put such a bench? Into `benches//bench/bevy_sprite/`?