@mockersf 2021-02-22
With `Visible` and `Transparent` as unit struct, how can one spawn an invisible or non-transparent `SpriteBundle`?