@rparrett 2021-02-23
That's a great question. I also asked in discord about manually managing visibility (but in a post-spawning scenario).