@rparrett 2021-02-23
Sprite rotation is also currently unaccounted for.