@Byteron 2021-02-23
> That's a great question. I also asked in discord about manually managing visibility (but in a post-spawning scenario).
> 
> https://discord.com/channels/691052431525675048/743663924229963868/813505768069398529
> 
> It sounds like Byteron's going to be looking into using a separate `Culled` marker struct instead, leaving the old `Visible` struct as-is.

> With `Visible` and `Transparent` as unit struct that are parts of the bundle, how can one spawn an invisible or non-transparent `SpriteBundle`?

Helloooo.
I reimplemented this with using a separate `Frustum` marker component.
Which means the `Visible` component stays untouched for this PR. There is also no need to manually add this Frustum Component yourself, that's done entirely by bevy.

> Sprite rotation is also currently unaccounted for.

This is still uncounted for. I need to figure out the math first in order to do this.