@Byteron 2021-02-23
> Could we use something like `WithinFrustum` that makes it more clear what the marker does? The name `Frustum` doesn't make it immediately obvious whether entities with the struct are inside or outside the camera frustum.

As I understand the Frustum is the visible part. So it makes sense to me that entities with Frustum are visible, while entities without Frustum are not.
Though if you think this is not descriptive enough I can rename it for sure.