@aevyrie 2021-02-23
A frustum is a shape, it doesn't really make much sense when we look at other component names:

- entities are `Visible`
- entities are `Transparent`
- entities are `Frustum` (?)

The first two make it clear what functionality the components add. Hence the suggestion for `WIthinFrustum`.