@Byteron 2021-02-23
Hm. I see your point.
Though I don't quite like the double word naming.
`Observable` maybe? `Ocular`?