@aevyrie 2021-02-23
> Hm. I see your point.
> Though I don't quite like the double word naming.
> `Observable` maybe? `Ocular`, `Viewable`?

Might need to leave this up to cart. Personally, I don't see a problem with multi-word names. Why trade clarity for one less capital letter?