@Byteron 2021-03-05
Updated after ECSv2 landed.
the `Viewable` component is now a `StorageType::SparseSet` as cart mentioned it should be earlier in this convo.
performance gains remain roughly the same.

The `StorageType::SparseSet` leads to slower iteration times in some rendering related systems (as they use `With<Viewable>`), so as soon as queries allow sparse set components to be in filters while still have fast iteration times there should be another small performance bump.