@rparrett 2021-03-17
I updated the example locally to non-uniformly scale the sprites
```rs
let scale = Vec3::new(0.2 + rng.gen::<f32>() * 4.8, 1.0, 1.0);
```
And I am seeing some popping-in and popping-out around the edges that I didn't notice before making that change.