@cart 2021-03-24
Just pushed the changes mentioned here: https://github.com/bevyengine/bevy/pull/1492#discussion_r600073824

This "inverts" the queries so that they search for `Without<OutsideFrustum>`. This allows non-sprite rendering to work by default without hacks (and notably without requiring SpritePlugin).