@cart 2021-03-24
Found the issue. It was a bevy_ecs bug introduced in the "reliable change detection" pr. Simple fix.