@cart 2021-03-24
An issue with using `OutsideFrustum` instead of `InsideFrustum` is that it won't support multiple views/cameras correctly. The old impl "under culled" objects viewed by multiple cameras. The new impl will "over cull" objects, which is much worse.  We don't support that sort of thing officially, but I think we should provide an escape hatch in case someone hits this issue in a custom render pipeline. Something like `SpriteSettings { frustum_culling_enabled: false }`.